requirements.txt
python-telegram-bot==21.10
python-dotenv==1.0.1
google-api-python-client==2.170.0
